Use subdir to control upload to IIIF S3 [WEB-1631]
This is difficult to explain. Essentially, when we upload to S3, we
have to specify a source folder, and everything inside that folder
will be uploaded to S3, with all of the relative paths kept intact.
Also, when we tell CloudFront to forward a request to S3, it will
keep the same path as it received.

We bound `iiif/static/*` to forward to S3. But we also want to have
flexibility about that path, so we can change it later if needed. So
we put a Lambda@Edge script between the CDN and S3, which forwards
requests from `iiif/static/*` to the root `/` of the S3 bucket. That
was the focus of the previous commit. This way, we don't need to
create "fake" `iiif/static` directories in S3 just to satisfy the
URL, and we don't need to create the directories locally, when
generating tile pyramids.

However, we can't just upload the entire `out` folder. If multiple
tile processes are running at the same time, each one would attempt
to upload _all_ of the pyramids inside that folder, not just the one
it was responsible for. Additionally, if this process fails for some
reason and doesn't clean up after itself, we'd be uploading all of
these half-broken pyramids on each run. We still need to put a
clean up script for that to run on schedule.

Creating a folder-within-a-folder allows us to target our upload at
the higher level directory, so that only the lower-level directory
gets uploaded. That's what this commit is about.

Also, I might be missing something, but it was surprisingly difficult
to find a way to remove a non-empty directory tree. I couldn't find a
way that was built-in to PHP or Flysystem. I'd have to write loops
that iterate over the tree manually. I suspect this task was made
difficult as a defensive measure.

All of the solutions I've found online say that they don't behave
quite like `rm -f`. I know that calling `exec` is always a risk, but
I think it'd be more risky to write my own functions to mass-remove
files, rather than just using a built-in one.

In any case, I added UUID validation to reduce the `exec` risk.

+++ app/Jobs/TileMedia.php
@@ -2,6 +2,8 @@
 
 namespace App\Jobs;
 
+use Exception;
+use Ramsey\Uuid\Uuid;
 use A17\Twill\Models\Media;
 use Illuminate\Support\Facades\Storage;
 
@@ -23,6 +25,10 @@ class TileMedia extends BaseJob
 
         $iiifMediaUuid = get_clean_media_uuid($iiifMedia);
 
+        if (!Uuid::isValid($iiifMediaUuid)) {
+            throw new Exception('Invalid UUID');
+        }
+
         $s3 = Storage::disk(config('twill.media_library.disk'));
         $local = Storage::disk('local');
         $iiifS3 = Storage::disk('iiif_s3');
@@ -30,13 +36,8 @@ class TileMedia extends BaseJob
         $localFilename = $iiifMediaUuid;
 
         // Clean up first if this job failed
-        if ($local->exists('tiles/src/' . $localFilename)) {
-            $local->delete('tiles/src/' . $localFilename);
-        }
-
-        if ($local->exists('tiles/out/' . $localFilename)) {
-            $local->deleteDirectory('tiles/out/' . $localFilename);
-        }
+        $this->deleteSrc($local, $localFilename);
+        $this->deleteOut($local, $localFilename);
 
         // No need to do all this work if the tiles have been generated..?
         if (!$this->forceRetile && $iiifS3->exists($localFilename)) {
@@ -56,11 +57,27 @@ class TileMedia extends BaseJob
         // There's probably a quicker way to transfer an entire directory to S3, but this'll do for now
         // https://bakerstreetsystems.com/blog/post/recursively-transfer-entire-directory-amazon-s3-laravel-52
         // https://stackoverflow.com/questions/44900585/aws-s3-copy-and-replicate-folder-in-laravel
+        // We are nesting a $localFilename directory inside a $localFilename directory and uploading the inner one
+        // This is done to prevent uploading the same directory twice, if multiple uploads are happening in parallel
         $s3Client = $iiifS3->getDriver()->getAdapter()->getClient();
-        $s3Client->uploadDirectory(storage_path() . '/app/tiles/out', config('filesystems.disks.iiif_s3.bucket'));
+        $s3Client->uploadDirectory(storage_path() . '/app/tiles/out/' . $localFilename, config('filesystems.disks.iiif_s3.bucket'));
 
         // Clean up on success
-        $local->delete('tiles/src/' . $localFilename);
-        $local->deleteDirectory('tiles/out/' . $localFilename);
+        $this->deleteSrc($local, $localFilename);
+        $this->deleteOut($local, $localFilename);
+    }
+
+    private function deleteSrc($local, $localFilename)
+    {
+        if ($local->exists('tiles/src/' . $localFilename)) {
+            $local->delete('tiles/src/' . $localFilename);
+        }
+    }
+
+    private function deleteOut($local, $localFilename)
+    {
+        if ($local->exists('tiles/out/' . $localFilename)) {
+            exec(sprintf("rm -r %s", escapeshellarg(storage_path() . '/app/tiles/out/' . $localFilename)));
+        }
     }
 }
